+++ ui/src/hooks/confirmSignUp/useComfirmSignUp.ts
@@ -20,9 +20,7 @@ export const useConfirmSignUp = () => {
         setErrorMessage('確認コードの有効期限が切れています。');
         break;
       default:
-        setErrorMessage(
-          'エラーが発生しました。しばらくしてからもう一度お試しください。'
-        );
+        setErrorMessage('エラーが発生しました。入力内容を確認してください。');
         break;
     }
   };
