+++ ui/src/pages/signup/SignUp.css
@@ -1,0 +1,4 @@
+.errorMessage {
+  color: red;
+  white-space: pre-wrap;
+}
